Add copyright, move Failure, rename files
