+++ src/argprocessing.hpp
@@ -19,6 +19,7 @@
 #pragma once
 
 #include "Args.hpp"
+#include "Statistic.hpp"
 
 #include "third_party/nonstd/optional.hpp"
